+++ src/d2_dll.h
