+++ src/d2_patch_const.h
