+++ src/d2_ptrs.h
