+++ src/dll_error.c
@@ -33,29 +33,39 @@
 *                                                                           *
 *****************************************************************************/
 
-#ifndef D2TEMPLATE89_PATCH_H_
-#define D2TEMPLATE89_PATCH_H_
+#include "dll_error.h"
 
 #include <stddef.h>
+#include <stdlib.h>
+#include <wchar.h>
+#include <windows.h>
 
-#include "d2_dll.h"
+#define MESSAGE_FORMAT L"File: %ls\n" \
+    L"Line: %d" \
+    L"%ls"
 
-#ifdef __cplusplus
-extern "C" {
-#endif /* __cplusplus */
+/*
+ * This is allocated as a global, so calls to error functions do not
+ * trigger stack overflow error.
+ */
+static wchar_t full_message[D2TEMPLATE_Error_kMessageCapacity];
 
-struct Patch {
-  enum D2Dll dll;
-  ptrdiff_t offset;
-  const void* data;
-  int is_relative;
-  size_t data_size;
-};
+/**
+ * External
+ */
 
-void Patch_Apply(struct Patch* patch);
+void D2TEMPLATE_ExitWithMessage(
+    const wchar_t* message,
+    const wchar_t* file,
+    int line) {
 
-#ifdef __cplusplus
-} /* extern "C" */
-#endif /* __cplusplus */
-
-#endif /* D2TEMPLATE89_PATCH_H_ */
+  _snwprintf(
+      full_message,
+      D2TEMPLATE_Error_kMessageCapacity,
+      MESSAGE_FORMAT,
+      file,
+      line,
+      message);
+  MessageBoxW(NULL, L"D2Template Error", message, MB_OK | MB_ICONERROR);
+  exit(EXIT_FAILURE);
+}
